Change the build process of the book.
The book can now have auto generated markdown files. This requires to have a
`build-src` temporary directory where all the markdown files are put in
the correct path before mdbook is ran. The include directory is a cache
of the auto generated markdown files that are either included in the markdown
step or html files patched in the final book.